Shuffle things around in preparation for Jest runner.
