+++ internal/runner/detector.go
@@ -2,11 +2,21 @@ package runner
 
 import (
 	"errors"
+	"os/exec"
 
 	"github.com/buildkite/test-splitter/internal/config"
+	"github.com/buildkite/test-splitter/internal/plan"
 )
 
-func DetectRunner(cfg config.Config) (*Rspec, error) {
+type TestRunner interface {
+	Command(testCases []string) (*exec.Cmd, error)
+	GetExamples(files []string) ([]plan.TestCase, error)
+	GetFiles() ([]string, error)
+	RetryCommand() (*exec.Cmd, error)
+	Name() string
+}
+
+func DetectRunner(cfg config.Config) (TestRunner, error) {
 	switch cfg.TestRunner {
 	case "rspec":
 		return NewRspec(Rspec{
